workaround for switching page transitions
Just overriding a CrossFade with a PageSlide leaves the target View unrendered for some reason.
Now using a CompositePageTransition with CompositePageTransition.PageTransitions[0] always set to either CrossFade or a dummy PageTransition (a modified CrossFade that won't change opacity). When adding a PageSlide, the effect will be visually replaced.

+++ GasPrices/Converters/ListViewCalculateMaxHeightConverter.cs
@@ -8,7 +8,7 @@ public class ListViewCalculateMaxHeightConverter : IValueConverter
 {
     public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
     {
-        return (double)value! * double.Parse(parameter!.ToString()!, CultureInfo.InvariantCulture);
+        return (int)((int)value! * double.Parse(parameter!.ToString()!, CultureInfo.InvariantCulture));
     }
 
     public object ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
